+++ src/fl_encoding_latin1.cxx
@@ -69,7 +69,7 @@ static int n_buf = 0;
 
 const char *fl_latin1_to_local(const char *t, int n)  
 {
-  if (n==-1) n = strlen(t);
+  if (n==-1) n = static_cast<int>(strlen(t));
   if (n<=n_buf) {
     n_buf = (n + 257) & 0x7fffff00;
     if (buf) free(buf);
@@ -90,7 +90,7 @@ const char *fl_latin1_to_local(const char *t, int n)
 
 const char *fl_local_to_latin1(const char *t, int n)
 {
-  if (n==-1) n = strlen(t);
+  if (n==-1) n = static_cast<int>(strlen(t));
   if (n<=n_buf) {
     n_buf = (n + 257) & 0x7fffff00;
     if (buf) free(buf);
